Upgrade to Fedora 38
Fedora 38 deprecated the docker daemon, so we move to using podman
instead. Both podman-docker and podman-remote are installed to enable
using remote docker commands with it.

Even though k3s should be tricked into running with podman, I've run
into some trouble with it, so for now it is only installed in the ubuntu
VM.